+++ Irvin.Fludal.SqlClient/SqlServer.cs
@@ -1,5 +1,4 @@
-﻿using System.Data;
-using System.Data.SqlClient;
+﻿using System.Data.SqlClient;
 
 namespace Irvin.Fludal.SqlClient;
